Film index works, database an be migrated

+++ Filminurk/Filminurk/Models/Movies/MoviesIndexViewModel.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string Title { get; set; }
         public DateOnly FirstPublished { get; set; }
-        public decimal? CurrentRating { get; set; }
+        public double? CurrentRating { get; set; }
 
         /* 3 õpilase valitud andmetüüpi */
         public string? Genre { get; set; }
